# Fast code path