--- a/packages/react/src/injectable-context.tsx
+++ b/packages/react/src/injectable-context.tsx
@@ -17,3 +17,10 @@ export function useInstance<T>(provider: Provider<T>): T {
 
   return childInjector.getInstance(provider)
 }
+
+export function useServerInstance<T>(provider: Provider<T>): T {
+  const childInjector = useContext(_InjectableContext)
+
+  // @ts-expect-error
+  return childInjector.serverCache.get(provider)
+}