… packages/react/src/index.browser.ts