### Before this optimize

```
┌─────────┬────────┬────────┬────────┬────────┬───────────┬─────────┬────────┐
│ Stat    │ 2.5%   │ 50%    │ 97.5%  │ 99%    │ Avg       │ Stdev   │ Max    │
├─────────┼────────┼────────┼────────┼────────┼───────────┼─────────┼────────┤
│ Latency │ 222 ms │ 245 ms │ 491 ms │ 536 ms │ 261.07 ms │ 60.5 ms │ 630 ms │
└─────────┴────────┴────────┴────────┴────────┴───────────┴─────────┴────────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Req/Sec   │ 57      │ 57      │ 77      │ 85      │ 76.24   │ 5.95    │ 57      │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Bytes/Sec │ 14.6 MB │ 14.6 MB │ 19.7 MB │ 21.8 MB │ 19.5 MB │ 1.52 MB │ 14.6 MB │
└───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┘
```

### After this optimize

```
┌─────────┬────────┬────────┬────────┬────────┬───────────┬──────────┬────────┐
│ Stat    │ 2.5%   │ 50%    │ 97.5%  │ 99%    │ Avg       │ Stdev    │ Max    │
├─────────┼────────┼────────┼────────┼────────┼───────────┼──────────┼────────┤
│ Latency │ 176 ms │ 216 ms │ 432 ms │ 464 ms │ 229.33 ms │ 51.91 ms │ 596 ms │
└─────────┴────────┴────────┴────────┴────────┴───────────┴──────────┴────────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Req/Sec   │ 60      │ 70      │ 88      │ 101     │ 87.02   │ 6.8     │ 60      │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Bytes/Sec │ 15.4 MB │ 17.9 MB │ 22.6 MB │ 25.9 MB │ 22.3 MB │ 1.74 MB │ 15.4 MB │
└───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┘
```

### After this and https://github.com/sigi-framework/sigi/pull/745

```
┌─────────┬────────┬────────┬────────┬────────┬───────────┬──────────┬────────┐
│ Stat    │ 2.5%   │ 50%    │ 97.5%  │ 99%    │ Avg       │ Stdev    │ Max    │
├─────────┼────────┼────────┼────────┼────────┼───────────┼──────────┼────────┤
│ Latency │ 196 ms │ 211 ms │ 420 ms │ 440 ms │ 222.57 ms │ 47.26 ms │ 549 ms │
└─────────┴────────┴────────┴────────┴────────┴───────────┴──────────┴────────┘
┌───────────┬─────────┬───────┬─────────┬─────────┬───────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%  │ 50%     │ 97.5%   │ Avg   │ Stdev   │ Min     │
├───────────┼─────────┼───────┼─────────┼─────────┼───────┼─────────┼─────────┤
│ Req/Sec   │ 65      │ 78    │ 90      │ 96      │ 89.67 │ 5.11    │ 65      │
├───────────┼─────────┼───────┼─────────┼─────────┼───────┼─────────┼─────────┤
│ Bytes/Sec │ 16.7 MB │ 20 MB │ 23.1 MB │ 24.6 MB │ 23 MB │ 1.31 MB │ 16.7 MB │
└───────────┴─────────┴───────┴─────────┴─────────┴───────┴─────────┴─────────┘
```